prevent deleting recipes from any email using subjects

+++ src/app/recipes/recipe-details/recipe-details.component.ts
@@ -1,7 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { RecipeService } from '../recipe.service';
 import { Recipe } from '../recipe.model';
+import { AuthService } from 'src/app/auth/auth.service';
+import { map } from 'rxjs/operators';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-recipe-details',
@@ -12,10 +15,13 @@ export class RecipeDetailsComponent implements OnInit {
   constructor(
     private recipeService: RecipeService,
     private route: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private authService: AuthService
   ) {}
+  private subscription: Subscription;
   recipe: Recipe;
   id: number;
+  deleteError: string = '';
   ngOnInit(): void {
     const id = this.route.params.subscribe((params: Params) => {
       this.id = Number(params['id']);
@@ -30,7 +36,20 @@ export class RecipeDetailsComponent implements OnInit {
     // this.router.navigate(['../', this.id, 'edit'], {relativeTo: this.route})
   }
   onDeleteRecipe() {
-    this.recipeService.deleteRecipe(this.id);
-    this.router.navigate(['../']);
+    this.subscription = this.authService.user.subscribe(
+      userData => {
+        if (userData.email === null) {
+          return;
+        }
+        if (userData.email === 'bishogeorge689@gmail.com') {
+          this.recipeService.deleteRecipe(this.id);
+          this.router.navigate(['../']);
+        }
+        else {
+          this.deleteError = 'you cannot delete the recipe please call the support';
+        }
+      }
+    )
   }
+
 }
